dps-spark/requirements.txt
kafka_python==2.0.2
pyspark==3.2.1